third_party/blink/web_tests/platform/generic/fast/files/file-reader-methods-illegal-arguments-expected.txt

Test FileReader methods and their argument type checking.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fileReader.readAsArrayBuffer(); threw exception TypeError: Failed to execute 'readAsArrayBuffer' on 'FileReader': 1 argument required, but only 0 present..
PASS fileReader.readAsArrayBuffer(2); threw exception TypeError: Failed to execute 'readAsArrayBuffer' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsArrayBuffer(null) threw exception TypeError: Failed to execute 'readAsArrayBuffer' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsArrayBuffer(fileReader) threw exception TypeError: Failed to execute 'readAsArrayBuffer' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsArrayBuffer([]) threw exception TypeError: Failed to execute 'readAsArrayBuffer' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsArrayBuffer(blob, 'utf-8') did not throw exception.
PASS fileReader.readAsArrayBuffer(blob) threw exception InvalidStateError: Failed to execute 'readAsArrayBuffer' on 'FileReader': The object is already busy reading Blobs..
PASS fileReader.readAsBinaryString(); threw exception TypeError: Failed to execute 'readAsBinaryString' on 'FileReader': 1 argument required, but only 0 present..
PASS fileReader.readAsBinaryString(2); threw exception TypeError: Failed to execute 'readAsBinaryString' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsBinaryString(null) threw exception TypeError: Failed to execute 'readAsBinaryString' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsBinaryString(fileReader) threw exception TypeError: Failed to execute 'readAsBinaryString' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsBinaryString([]) threw exception TypeError: Failed to execute 'readAsBinaryString' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsBinaryString(blob, 'utf-8') did not throw exception.
PASS fileReader.readAsBinaryString(blob) threw exception InvalidStateError: Failed to execute 'readAsBinaryString' on 'FileReader': The object is already busy reading Blobs..
PASS fileReader.readAsText(); threw exception TypeError: Failed to execute 'readAsText' on 'FileReader': 1 argument required, but only 0 present..
PASS fileReader.readAsText(2); threw exception TypeError: Failed to execute 'readAsText' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsText(null) threw exception TypeError: Failed to execute 'readAsText' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsText(fileReader) threw exception TypeError: Failed to execute 'readAsText' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsText([]) threw exception TypeError: Failed to execute 'readAsText' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsText(blob, 'utf-8') did not throw exception.
PASS fileReader.readAsText(blob) threw exception InvalidStateError: Failed to execute 'readAsText' on 'FileReader': The object is already busy reading Blobs..
PASS fileReader.readAsDataURL(); threw exception TypeError: Failed to execute 'readAsDataURL' on 'FileReader': 1 argument required, but only 0 present..
PASS fileReader.readAsDataURL(2); threw exception TypeError: Failed to execute 'readAsDataURL' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsDataURL(null) threw exception TypeError: Failed to execute 'readAsDataURL' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsDataURL(fileReader) threw exception TypeError: Failed to execute 'readAsDataURL' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsDataURL([]) threw exception TypeError: Failed to execute 'readAsDataURL' on 'FileReader': parameter 1 is not of type 'Blob'..
PASS fileReader.readAsDataURL(blob, 'utf-8') did not throw exception.
PASS fileReader.readAsDataURL(blob) threw exception InvalidStateError: Failed to execute 'readAsDataURL' on 'FileReader': The object is already busy reading Blobs..
PASS successfullyParsed is true

TEST COMPLETE
